aftr-python/requirements.txt
fastapi
jose
passlib
sqlalchemy
dotenv
pydantic[email]
pillow
